Don't open hero social link if href is local
In case link for a social icon in the hero image is set to local link
(e.g., "#contact"), don't open it in a new tab. It's useful if you don't
want to publish a link to your e-mail in clear text, and rather redirect
user to the contact form.

Signed-off-by: Marek Pikuła <marek@serenitycode.dev>